+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When the campus API fails to return a floor plan, the building block now returns an empty floor plan object instead of a 500 status
 ### added
 - Add section number to course data for MyCourses
+- Only display floor plans for floors that are in FAM campus web application
 
 ## [2.11.1] - 2024-11-04
 ### Changed